fix(returns): adversarial-review fixes — token mint for email-less orders, pending-aware budgets, portal/tracking hygiene
CONFIRMED: maybeSendReturnPrompt (the ONLY return_token mint site) bailed on
no_requester_email BEFORE minting, so staff-created internal orders
(requester_email NULL by construction) never got a token and their requester
could never see the dashboard 'Request a return' link. Reordered: the guarded
mint (.is('return_token', null) — never rotates) now runs for every
structurally-qualifying completion (completed + returns module on +
fulfilled > 0); requester_email + the 0278 marker gate only the EMAIL
(marker = email sent, untouched on an email-less completion).

Minors:
- portalOrders per-order returns list now .eq('source','requester') — staff
  internal returns (incl. denied/cancelled) no longer leak onto the portal.
- .int() on return quantity schemas (v1 route, createLineSchema,
  requesterLineSchema) — portal already enforced it.
- App-side 'remaining' now subtracts live PENDING (unapplied, not
  cancelled/denied) return_lines, mirroring the 0153 DB cap trigger, via a
  shared pendingReturnQuantitiesByLine: requester context (token + portal),
  staff createFromOrder validation, web returnableLinesForOrder, and the
  portal returnable calc (new quantityPendingReturn on PortalOrder lines).
  Mobile client math left as-is (server re-validates via the v1 route).
- Staff createFromOrder now rolls back the orphan 'requested' header when the
  lines insert fails — through the service-role client, since returns RLS
  deliberately denies DELETE to user-authed clients (0153/0154).
- Tracking-page returnPath now treats legacy 'delivered' as terminal, same as
  every other returns surface.

Tests: mint-for-email-less-order, pending-budget (staff + token + portal),
cancelled/denied-don't-count, fractional-quantity rejects, admin-client
rollback. Web suite 283 files / 2456 tests green; typecheck + eslint clean.

+++ apps/web/src/app/api/v1/orders/[id]/returns/route.ts
@@ -28,7 +28,9 @@ const bodySchema = z.object({
     .array(
       z.object({
         orderRequestLineId: z.string().uuid(),
-        quantity: z.coerce.number().positive(),
+        // Whole units only — fractional return quantities are data-shape
+        // nonsense (the DB cap trigger backstops over-return, not shape).
+        quantity: z.coerce.number().int().positive(),
         disposition: z.enum(['restock', 'scrap']),
       }),
     )
